+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 - Fixed formatting of structures after a label.
 - Visibility specifiers (e.g. `private`, `public`) are no longer treated like section headers when
   used outside type declarations.
+- Fixed formatting of comments before the first item in import and exports clauses.
 
 ## [0.4.0] - 2025-03-18